decouple running jobs process from express app process, change launch commands
