+++ package.json
@@ -3,7 +3,9 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "nodejs ./bin/www"
+    "start": "./node_modules/.bin/forever start launch.json",
+    "stop": "./node_modules/.bin/forever stop express; ./node_modules/.bin/forever stop cron; exit 0;",
+    "restart": "./node_modules/.bin/forever restart express; ./node_modules/.bin/forever restart cron; exit 0;"
   },
   "dependencies": {
     "async": "^1.5.2",
@@ -12,6 +14,7 @@
     "cron": "^1.1.0",
     "debug": "~0.7.4",
     "express": "~4.0.0",
+    "forever": "^0.15.1",
     "jade": "~1.3.0",
     "mandrill-api": "^1.0.45",
     "moment": "^2.11.0",
